+++ MOP/src/Common/ExceptionManager.cs
@@ -43,7 +43,7 @@ namespace MOP.Common
             if (erorrsContainer.Contains(message))
                 return;
 
-            string fileName = $"{DefaultErrorLogName}_{DateTime.Now.ToString("yyyy-MM-dd-HH-mm")}";
+            string fileName = $"{DefaultErrorLogName}_{DateTime.Now:yyyy-MM-dd-HH-mm}";
 
             if (File.Exists($"{GetLogFolder()}/{fileName}.txt"))
             {
@@ -168,12 +168,21 @@ namespace MOP.Common
             // Game data
             if (ModLoader.GetCurrentScene() == CurrentScene.Game)
             {
-                output += "\n=== GAME DATA ===\n\n";
-                output += $"PlayerPosition: {GameObject.Find("PLAYER").transform.position}\n";
-                output += $"PlayerHasHayosikoKey: {FSM.FsmManager.PlayerHasHayosikoKey()}\n";
-                output += $"IsPlayerInCar: {FSM.FsmManager.IsPlayerInCar()}\n";
-                output += $"IsPlayerInSatsuma: {FSM.FsmManager.IsPlayerInSatsuma()}\n";
-                output += $"CanTriggerStatus: {(Managers.ItemsManager.Instance.GetCanTrigger() == null ? "null" : $"Found ({Managers.ItemsManager.Instance.GetCanTrigger().GetGameObjectPath()})")}\n";
+                try
+                {
+                    output += "\n=== GAME DATA ===\n\n";
+                    output += $"PlayerPosition: {GameObject.Find("PLAYER").transform.position}\n";
+                    output += $"PlayerHasHayosikoKey: {FSM.FsmManager.PlayerHasHayosikoKey()}\n";
+                    output += $"IsPlayerInCar: {FSM.FsmManager.IsPlayerInCar()}\n";
+                    output += $"IsPlayerInSatsuma: {FSM.FsmManager.IsPlayerInSatsuma()}\n";
+                    output += $"DrawDistance: {FSM.FsmManager.GetDrawDistance()}\n";
+                    output += $"CanTriggerStatus: {(Managers.ItemsManager.Instance == null ? "manager_null" : Managers.ItemsManager.Instance.GetCanTrigger() == null ? "null" : $"Found ({Managers.ItemsManager.Instance.GetCanTrigger().GetGameObjectPath()})")}\n";
+                    output += $"IsTrailerAttached: {FSM.FsmManager.IsTrailerAttached()}\n";
+                }
+                catch
+                {
+                    output += "ERROR GETTING GAME DATA!";
+                }
             }
 
             // List installed mods.
@@ -184,7 +193,7 @@ namespace MOP.Common
                 if (mod.ID.EqualsAny("MSCLoader_Console", "MSCLoader_Settings", "MOP"))
                     continue;
 
-                output += $"{mod.Name}:\n  ID: {mod.ID}\n  Version: {mod.Version}\n  Author: {mod.Author}\n\n";
+                output += $"{mod.Name}:\n  ID: {mod.ID}\n  Version: {mod.Version}\n  Author: {mod.Author}\n  IsDisabled: {mod.isDisabled}\n\n";
             }
 
             // If only 3 mods have been found, that means the only mods active are MOP and two ModLoader modules.