+++ MOP/src/Hypervisor.cs
@@ -101,11 +101,6 @@ namespace MOP
             cursorFSM = GameObject.Find("PLAYER").GetPlayMakerByName("Update Cursor");
             cursorFSM.enabled = false;
 
-            // Start the delayed initialization routine
-            PlayMakerFSM[] gtGrille = GameObject.Find("grille gt(Clone)").GetComponents<PlayMakerFSM>();
-            foreach (var fsm in gtGrille)
-                fsm.Fsm.RestartOnEnable = false;
-
             // Disable rule files if user wants it.
             if (!RulesManager.Instance.LoadRules)
             {
@@ -119,7 +114,9 @@ namespace MOP
         #region MOP Initialization
         IEnumerator DelayedInitializaitonRoutine()
         {
-            yield return new WaitForSeconds(2);
+            //yield return new WaitForSeconds(2);
+            for (int i = 0; i < 5; i++)
+                yield return null;
             Initialize();
         }
 
@@ -171,7 +168,6 @@ namespace MOP
                 ExceptionManager.New(ex, false, "WORLD_OBJECTS_1_INITIALIZAITON_FAIL");
             }
 
-
             // Initialize places.
             placeManager = new PlaceManager();
 
@@ -205,7 +201,6 @@ namespace MOP
                 worldObjectManager.Add("SwampColliders", DisableOn.PlayerInHome);
                 worldObjectManager.Add("RYKIPOHJA", DisableOn.PlayerInHome);
                 worldObjectManager.Add("COMPUTER", DisableOn.PlayerAwayFromHome);
-                //worldObjectManager.Add("JOBS/HouseDrunkNew", DisableOn.PlayerInHome);
 
                 ModConsole.Print("[MOP] World objects (2) loaded");
             }
@@ -1136,6 +1131,12 @@ namespace MOP
             catch (Exception ex)
             {
                 ExceptionManager.New(ex, false, "TOGGLE_ALL_SATSUMA_TOGGLE_ELEMENTS");
+            }
+
+            // Teleport real radiatorHose3 to currentRadiatorHose.
+            if (mode == ToggleAllMode.OnSave)
+            {
+                ItemsManager.Instance.TeleportRealRadiatorHose();
             }
         }